Fix compiler warnings about %ld versus time_t on systems where time_t is a synonym for long long instead of long.

+++ agent/mibgroup/disman/schedule/schedCore.c
@@ -301,16 +301,16 @@ sched_nextTime( struct schedTable_entry *entry )
         } else {
              entry->schedNextRun = now + entry->schedInterval;
         }
-        DEBUGMSGTL(("disman:schedule:time", "periodic: (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+        DEBUGMSGTL(("disman:schedule:time", "periodic: (%lld) %s",
+                    (long long) entry->schedNextRun,
+                    ctime(&entry->schedNextRun)));
         break;
 
     case SCHED_TYPE_ONESHOT:
         if ( entry->schedLastRun ) {
-            DEBUGMSGTL(("disman:schedule:time", "one-shot: expired (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+            DEBUGMSGTL(("disman:schedule:time", "one-shot: expired (%lld) %s",
+                        (long long) entry->schedNextRun,
+                        ctime(&entry->schedNextRun)));
             return;
         }
         /* Fallthrough */
@@ -424,9 +424,9 @@ sched_nextTime( struct schedTable_entry *entry )
          * 'next_tm' now contains the time for the next scheduled run
          */
         entry->schedNextRun = mktime( &next_tm );
-        DEBUGMSGTL(("disman:schedule:time", "calendar: (%ld) %s",
-                                  entry->schedNextRun,
-                           ctime(&entry->schedNextRun)));
+        DEBUGMSGTL(("disman:schedule:time", "calendar: (%lld) %s",
+                    (long long) entry->schedNextRun,
+                    ctime(&entry->schedNextRun)));
         return;
 
     default:
